Добавляет обработку ошибки в случае недоступности сервера
