+++ src/model/points-model.js
@@ -25,6 +25,7 @@ export default class PointsModel extends Observable {
       this.#points = points.map(this.#adaptToClient);
     } catch(err) {
       this.#points = [];
+      this._notify(UpdateType.ERROR);
     }
     this._notify(UpdateType.INIT);
   }
