+++ src/view/no-point-view.js
@@ -1,12 +1,5 @@
 import AbstractView from '../framework/view/abstract-view.js';
-import {FilterType} from '../utils/consts.js';
-
-const NoPointsTextType = {
-  [FilterType.EVERYTHING]: 'Click New Event to create your first point',
-  [FilterType.FUTURE]: 'There are no future events now',
-  [FilterType.PRESENT]: 'There are no present events now',
-  [FilterType.PAST]: 'There are no past events now',
-};
+import {NoPointsTextType} from '../utils/consts.js';
 
 function createNoPointTemplate(filterType) {
   const noPointTextValue = NoPointsTextType[filterType];
